+++ DIBBLES/Systems/Time.cs
@@ -4,15 +4,15 @@ namespace DIBBLES.Systems;
 
 public class Time
 {
-    public static float time;
+    public static double time;
     
-    private static float deltaTime;
-    public static float DeltaTime
+    private static double deltaTime;
+    public static double DeltaTime
     {
         get
         {
             // Clamp deltaTime to prevent teleportation when lagging
-            return !Engine.IsPaused ? GMath.Clamp(deltaTime, 0f, 0.1f) : 0f;
+            return !Engine.IsPaused ? GMath.Clamp(deltaTime, 0d, 0.1d) : 0d;
         }
         set{ deltaTime = value; }
     }
